refactor(agents): the identity results move out of results.go
It crossed the 500-line cap. The three that moved answer WHO and
WHAT-IT-IS-CALLED rather than what a record holds — the acting human, the
colleague roster, the tag vocabulary — and each answers a question a caller has
to settle BEFORE it can name a record: whose id goes in owner_id, which
colleague gets the task, which word to tag with.
